Support passing an autosign token as part of the build
If using the autosign gem to generate one time or time limited tokens
you may want to pass them the agent process when building with a Puppet
master.

The implementation removes the token from the image so it's not
available in any of the layers. It's also not saved in the Dockerfile,
rather it's passed in as a build-arg.