Support acbuild for creating rkt compatible images
This commit adds a new face for building aci images for use with rkt.
This involves a bunch of refacctoring to move some of the docker
specific code out and make a generic image building think. This isn't
quite all done, but enough for the purposes of adding rkt support.

This goes hand-in-hand with changing the name from docker_build to
image_build, to represent the wider remit.