Resolve issue with passing through arguments
Two bugs fixed here. One relating to passing through arguments. Because
Puppet automatically casts hyphens to underscores the pass-through args
didn't work as expected. On top of that anything starting with no
automatically gets set to false, which is not what we want for the
no-cache pass-through. So we catch the special case and set it to true.