Check return code of puppet apply.
Currently the docker build succeeds even if puppet apply does not
succeed.

Puppet apply has a non-standrad system of return codes. It will
return 0 even if it hasn't sucessfully applied the manifest unless
`--detailed-exitcodes` is given. When this is given if it returns
0 or 2 the manifest has been applied sucessfully [0].

This patch makes the docker build fail if the manifest is not
applied successfully.

[0] https://docs.puppet.com/puppet/latest/man/apply.html#OPTIONS