Allow specifying of facter variables during puppet apply.
Sometimes you want to push additional variables to facter during
puppet apply. For instance, secrets should not be stored under version
control but are needed when the configuration is applied. In this instance
using passing the secret as a facter variable to puppet is a nice way
of having the configured machine have the secret but not include the secret
in your source code.

This patch lets you specify facter variables at parameters to
`puppet docker build` that are in turn sent to the `puppet apply` call.