Add support for puppet 6
This commit adds support for puppet 6 to the image_build module. Without this change the module will only install puppet 5, even if puppet version 6 is specified. Additionally, this change removes the app_management parameter from Puppet since it is no longer used or supported in puppet 6. Without this change image build will not run with, or install, puppet version 6.